measureForX either runDuration as cycles or seconds.
On mac laptop there really doesn't seem to be any difference. BUT I'm
including the mac binaries, the weird this is bmc and bmt were compiled
before I switched the bmRunX suite run to be 2 minutes and both of these
measure the loop time as 26 cycles. But the bmcl and bmct where measure
it at 56 cycles! Why?????