Add bmsuite.nim
A hack, but it does work.

I want to be able to use benchmark.suite inside a proc, but since its
public, i.e. its declared with an asterick, a compile error
occurs indicating '*' not allowed. So I moved it into a separate file,
bmsuite.nim which can be included and then invoked with bmSuite. You'll
also need to 'import benchmark'.